M3-8898 move to separate summary file
